client/requirements.txt
# SurveillX Streaming Client Requirements

# WebSocket client for real-time communication
python-socketio[client]>=5.0.0
websocket-client>=1.3.0

# Video capture and processing
opencv-python>=4.5.0

# HTTP requests
requests>=2.25.0
